Good to know; php match function is checking on type as well. date() always returns a string so not gonna work without type casting.

+++ app/ChatMessage.php
@@ -61,7 +61,7 @@ class ChatMessage
      */
     public static function getByAccountId(int $accountId, int $amount = 5): array
     {
-        return match (date('N')) {
+        return match ((int)date('N')) {
             4 => self::getThrowbackThursdayByAccountId($accountId, $amount),
             7 => self::getContextSundayByAccountId($accountId, 10),
             default => self::getRandomByAccountId($accountId, $amount),
